Put maximized windows on top
If we switch focus with the mouse, the newly-focussed window is not
raise() to the top. This doesn't make a difference in tiling mode, but
when we subsequently maximize the window, suddenly the focussed window
is not visible (because not on top).

This patch raises the focussed window before maximizing it.